+++ src/api/Compooler.API.Tests/Compooler.API.Tests.csproj
@@ -11,7 +11,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="CookieCrumble" Version="14.0.0-p.138" />
+        <PackageReference Include="CookieCrumble" Version="14.0.0-rc.0" />
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
         <PackageReference Include="xunit" Version="2.9.0"/>
         <PackageReference Include="xunit.runner.visualstudio" Version="2.8.2">
