+++ src/api/Compooler.API/Compooler.API.csproj
@@ -13,8 +13,8 @@
     </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="HotChocolate.AspNetCore" Version="14.0.0-p.138" />
-      <PackageReference Include="HotChocolate.Data.EntityFramework" Version="14.0.0-p.138" />
+      <PackageReference Include="HotChocolate.AspNetCore" Version="14.0.0-rc.0" />
+      <PackageReference Include="HotChocolate.Data.EntityFramework" Version="14.0.0-rc.0" />
       <PackageReference Include="JetBrains.Annotations" Version="2024.2.0" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.7">
         <PrivateAssets>all</PrivateAssets>
