+++ src/api/Compooler.Persistence/DataLoaders/Entities/RideIdsByUserIdDataLoader.cs
@@ -1,11 +1,9 @@
-using Compooler.Persistence;
 using Compooler.Persistence.Configurations;
-using JetBrains.Annotations;
+using GreenDonut;
 using Microsoft.EntityFrameworkCore;
 
-namespace Compooler.API.DataLoaders.Entities;
+namespace Compooler.Persistence.DataLoaders.Entities;
 
-[UsedImplicitly]
 public class RideIdsByUserIdDataLoader(
     IServiceProvider serviceProvider,
     IBatchScheduler batchScheduler,
